Refactored CCLocationManager file
